* commit '3e41c45260287e6175d4811e6cd30d7b3e8313c0':
  Uncommented replacePlantsOnAllPlots in PlotDAOImp
  Werkt weer

+++ src/main/java/nl/han/oose/sapporo/resource/PlantResource.java
@@ -15,6 +15,8 @@ import javax.ws.rs.core.Response;
 public class PlantResource {
     private IPlantService plantService;
     private IAccountService accountService;
+    private IAdminService adminService;
+    private IPlotService plotService;
 
     @Inject
     public void setPlantService(IPlantService plantService) {
@@ -24,6 +26,16 @@ public class PlantResource {
     @Inject
     public void setAccountService(IAccountService accountService) {
         this.accountService = accountService;
+    }
+
+    @Inject
+    public void setAdminService(IAdminService adminService) {
+        this.adminService = adminService;
+    }
+
+    @Inject
+    public void setPlotService(IPlotService plotService) {
+        this.plotService = plotService;
     }
 
     @GET
@@ -40,6 +52,8 @@ public class PlantResource {
     @Produces(MediaType.APPLICATION_JSON)
     public Response deletePlant(@QueryParam("token") String token, @PathParam("plantIDToDelete") int plantIDToDelete, @PathParam("plantIDToReplaceWith") int plantIDToReplaceWith) {
         UserDTO user  = accountService.verifyToken(token);
+        adminService.checkIfUserIsAdmin(user);
+        plotService.replacePlantsOnAllPlots(plantIDToDelete, plantIDToReplaceWith);
         plantService.deletePlant(user, plantIDToDelete, plantIDToReplaceWith);
         return Response.status(Response.Status.OK)
                 .entity(plantService.getAllPlants())
